Batch 3: Architecture + Distribution improvements
Architecture:
14. initSharedConfig(): Single init call replaces 4 scattered setters.
    Desktop main.tsx now calls one function with typed config object.
15. SWR compare (done in Batch 2)
16. Split evaluation-service.ts: parse-tool-response.ts,
    statistical-evaluator.ts, evaluation-logger.ts. Original file
    becomes barrel re-export for backward compatibility.
17. Move choice-tracker to features/connection/use-choice-tracker.ts
    (was in evaluation/ but it's a React hook with useRef/"use client")

Distribution:
11. Code signing placeholders: macOS signingIdentity + providerShortName,
    Windows certificateThumbprint + digestAlgorithm + timestampUrl
12. tauri-plugin-updater: Added dependency, registered plugin,
    createUpdaterArtifacts enabled in bundle config
13. Binary header verification: DLL downloads checked for valid
    PE (MZ) or Mach-O header before installation
