+++ apps/desktop/src-tauri/Cargo.toml
@@ -21,6 +21,7 @@ serde = { version = "1.0", features = ["derive"] }
 log = "0.4"
 tauri = { version = "2.10.3", features = [] }
 tauri-plugin-log = "2"
+tauri-plugin-updater = "2"
 reqwest = { version = "0.12", features = ["json", "stream", "rustls-tls"] }
 tokio = { version = "1", features = ["fs", "io-util"] }
 zip = "2.0"
