+++ apps/desktop/src-tauri/capabilities/default.json
@@ -6,6 +6,7 @@
     "main"
   ],
   "permissions": [
-    "core:default"
+    "core:default",
+    "updater:default"
   ]
 }
