+++ apps/desktop/src/main.tsx
@@ -3,9 +3,9 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import { App } from "./App";
 import { AuthProvider } from "./auth-provider";
-import { setApiBaseUrl, setAccessTokenGetter } from "@sts2/shared/lib/api-client";
-import { initSupabase, createClient } from "@sts2/shared/supabase/client";
-import { setAuthRedirectOrigin } from "@sts2/shared/supabase/auth";
+import { initSharedConfig } from "@sts2/shared/lib/init";
+import { setApiBaseUrl } from "@sts2/shared/lib/api-client";
+import { createClient } from "@sts2/shared/supabase/client";
 
 // Configure shared package for desktop environment
 // TODO: Read from environment or Tauri config
@@ -13,18 +13,21 @@ const API_BASE = import.meta.env.VITE_API_BASE_URL ?? "https://sts2-helper.verce
 const SUPABASE_URL = import.meta.env.VITE_SUPABASE_URL ?? "";
 const SUPABASE_ANON_KEY = import.meta.env.VITE_SUPABASE_ANON_KEY ?? "";
 
-setApiBaseUrl(API_BASE);
-setAuthRedirectOrigin(API_BASE);
-
 if (SUPABASE_URL && SUPABASE_ANON_KEY) {
-  initSupabase(SUPABASE_URL, SUPABASE_ANON_KEY);
-
-  // Desktop app sends Bearer token for API auth (no cookies cross-origin)
-  setAccessTokenGetter(async () => {
-    const { data } = await createClient().auth.getSession();
-    return data.session?.access_token ?? null;
+  initSharedConfig({
+    supabaseUrl: SUPABASE_URL,
+    supabaseAnonKey: SUPABASE_ANON_KEY,
+    apiBaseUrl: API_BASE,
+    authRedirectOrigin: API_BASE,
+    // Desktop app sends Bearer token for API auth (no cookies cross-origin)
+    accessTokenGetter: async () => {
+      const { data } = await createClient().auth.getSession();
+      return data.session?.access_token ?? null;
+    },
   });
 } else {
+  // API base still needed for non-auth features
+  setApiBaseUrl(API_BASE);
   console.warn("[STS2] Supabase credentials not configured. Auth features will not work.");
 }
